Fixes ESP image path, use mtools and hardlink caching
The lorax builder creates the ESP image under different path, therefore
the part that mounts it and overwrites the grub config no longer works.

In addition to that, this bit is the only one that requires root, this
patch replaces this with mcopy from mtools package which operates on DOS
images directly without mounting. No root needed anymore.

Finaly, I have found out that the remastered images are bigger by 50MB.
It looks like the kernel and initramdisk are symlink on the original
ISO, however, when copyined to a temp directory they are regular files.
Although mkisofs detects hardlinks by default, it cannot find them as it
does not perform contents detection. The script now deletes a copy and
makes them hardlinks again so the remastered image is the same size as
the original saving 50MB of space.